# docs/mobile-openapi.yaml
openapi: 3.0.3
info:
  title: NGA Admin Mobile API (subset)
  version: "1.0.0"
  description: |
    Staff native clients use Bearer JWT on the same Next.js `/api/*` hosts as the web app.
    Optional header `X-NGA-Client-Version` identifies app builds for server logs.
servers:
  - url: https://{host}
    variables:
      host:
        default: example.com
...
paths:
...
  /api/bookings/upload:
    post:
      tags: [bookings]
      summary: Multipart document upload
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file, bookingId, type]
              properties:
                file: { type: string, format: binary }
                bookingId: { type: string }
                type: { type: string, enum: [id_document, insurance_proof] }
      responses:
        "200": { description: Stored }
        "400": { description: Type/size/MIME error }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...